+++ meta/src/robocorp/_meta/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "1.6.0"
+__version__ = "1.6.1"
 version_info = [int(x) for x in __version__.split(".")]