+++ windows/docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.2 - 2024-02-01
+
+- Notice in the docs about the need to include `robocorp-windows` dependency in order
+  to make it available in the environment. (not included by default in `robocorp`)
+
 ## 1.0.1 - 2024-01-14
 
 - Fix main README and update docs.